requirements.txt
# requirements.txt - Dépendances Python

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
#psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
pyjwt==2.8.0

# AI/ML
openai==1.3.7
faiss-cpu==1.7.4
numpy==1.24.3

# Document Processing
PyPDF2==3.0.1
python-pptx==0.6.23
python-docx==1.1.0
Pillow==10.1.0
pytesseract==0.3.10

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2

# Streamlit (Frontend)
streamlit==1.29.0

# Testing
pytest==7.4.3
httpx==0.25.2
